gitautosync update 75

+++ myprog.py
@@ -10,7 +10,7 @@ print(hello('world'))
 # print(hello(hello))
 
 
-def mysum(*args: int) -> int:
+def mysum(*args: float) -> float:
     total = 0
 
     for one_arg in args:
